@@ -8,9 +8,11 @@
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "date-fns": "^3.4.0",
     "axios": "^1.6.7",
     "notiflix": "^3.2.7",
-    "react": "^18.1.0",
+    "react": "^18.2.0",
+    "react-datepicker": "^6.3.0",
     "react-dom": "^18.1.0",
     "react-helmet-async": "^2.0.4",
     "react-loader-spinner": "^6.1.6",
